+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * add stratified analysis to `tm_g_forest_tte` and `tm_g_forest_rsp`
 
+* add confidence level and plotting symbol size options to `tm_g_forest_rsp` and `tm_g_forest_tte`
+
 # teal.modules.clinical 0.8.1
 
 * new `tm_t_events` and `tm_t_events_by_grade` modules